Generate haptics on input errors
